requirements.txt
accelerate
cssutils
beautifulsoup4
evaluate
pydantic
pydantic_settings
pymongo
duden
dict2xml
torch
transformers
scikit-learn
sklearn-crfsuite
pandas
numpy
seqeval
spacy<3.7.0
spacy_conll
tensorflow
tqdm
requests
matplotlib
scipy
seaborn